refactor(agents): move lifecycle context types out of the controller module to break the architecture cycle
